Create handler and continue refactoring
