+++ redis-database-service/application/brocker.go
@@ -1,0 +1,9 @@
+package application
+
+import (
+	mqtt "github.com/eclipse/paho.mqtt.golang"
+)
+
+var (
+	MqttClient *mqtt.Client
+)
