+++ redis-database-service/boot/brocker.go
@@ -1,9 +1,10 @@
 package boot
 
 import (
+	"redis-database-service/infrastructure/handler"
 	"fmt"
 	"log"
-	"redis-database-service/cmd/application"
+	"redis-database-service/application"
 	"time"
 
 	mqtt "github.com/eclipse/paho.mqtt.golang"
@@ -31,17 +32,9 @@ func Connect() {
 
 // ListenBrocker : Function écoutant le brocker et insérant les données reçues, dans la base de données
 func ListenBrocker() {
-	
 	var wg sync.WaitGroup
 	wg.Add(1)
-	fmt.Println("En attente de message...")
-	data := entities.Measure{}
-	MqttClient.Subscribe(boot.TopicName, 0, func(client mqtt.Client, msg mqtt.Message) {
-		// Parse le JSON dans un objet à chaque reception d'un message sur le topic
-		log.Println("Message reçu : " + string(msg.Payload()))
-		json.Unmarshal([]byte(msg.Payload()), &data)
-		// Ajoute la measure dans la base de données
-		go redis.Save(data)
-	})
+	fmt.Println("Waiting messages...")
+	MqttClient.Subscribe(boot.TopicName, 0, handler.MeasureHandler)
 	wg.Wait()
 }
