--- redis-database-service/cmd/brocker/connecting.go
@@ -1,29 +1,0 @@
-package brocker
-
-import (
-	"fmt"
-	"log"
-	"time"
-
-	mqtt "github.com/eclipse/paho.mqtt.golang"
-)
-
-func createClientOptions(brokerURI string, clientID string) *mqtt.ClientOptions {
-	opts := mqtt.NewClientOptions()
-	opts.AddBroker(brokerURI)
-	opts.SetClientID(clientID)
-	return opts
-}
-
-func Connect(brokerURI string, clientID string) mqtt.Client {
-	fmt.Println("Trying to connect (" + brokerURI + ", " + clientID + ")...")
-	opts := createClientOptions(brokerURI, clientID)
-	client := mqtt.NewClient(opts)
-	token := client.Connect()
-	for !token.WaitTimeout(3 * time.Second) {
-	}
-	if err := token.Error(); err != nil {
-		log.Fatal(err)
-	}
-	return client
-}
