+++ redis-database-service/infrastructure/repository/measure.go
@@ -3,8 +3,8 @@ package repository
 import (
 	"encoding/json"
 	"fmt"
-	"redis-database-service/cmd/application"
-	entities "redis-database-service/internal/entities"
+	"redis-database-service/application"
+	"redis-database-service/domain"
 	"strconv"
 
 	redis "github.com/go-redis/redis/v7"
@@ -18,7 +18,8 @@ func NewMeasureRepository() *MeasureRepository {
 	return &MeasureRepository{}
 }
 
-func (m *MeasureRepository) SaveMeasure() error {
+// SaveMeasure save all the data in the database
+func (m *MeasureRepository) SaveMeasure(measure domain.Measure) error {
 	saveMeasureJSON(measure)
 	saveMeasureTimestamp(measure)
 	saveMeasureValue(measure)
@@ -26,7 +27,7 @@ func (m *MeasureRepository) SaveMeasure() error {
 
 func saveMeasureValue(measure domain.Measure) error {
 	// TODO : Prendre en compte le nombre de ligne retourné et mettre une erreur en conséquence
-	application.RedisClient.Zadd(getKeyZaddValue(measure), &redis.Z{
+	application.RedisClient.ZAdd(getKeyZaddValue(measure), &redis.Z{
 		Score:  float64(measure.MeasureValue),
 		Member: getKeySet(measure),
 	})
@@ -41,7 +42,7 @@ func saveMeasureTimestamp(measure domain.Measure) error {
 }
 
 func saveMeasureJSON(measure domain.Measure) error {
-	dataJSON, err := json.Marshal(data)
+	dataJSON, err := json.Marshal(measure)
 	if err != nil {
 		return errors.Wrap(err, "error saveMeasureJSON:json.Marshal")
 	}
@@ -49,18 +50,18 @@ func saveMeasureJSON(measure domain.Measure) error {
 	application.RedisClient.Set(getKeySet(measure), fmt.Sprintf("%s", dataJSON), 0)
 }
 
-func getKeySet(measure entities.Measure) string {
+func getKeySet(measure domain.Measure) string {
 	key := fmt.Sprintf("sensor:%s:measure:%s", strconv.Itoa(measure.SensorID), strconv.Itoa(measure.Timestamp))
 	fmt.Println("Key : " + key)
 	return key
 }
 
-func getKeyZaddValue(measure entities.Measure) string {
+func getKeyZaddValue(measure domain.Measure) string {
 	key := fmt.Sprintf("measure_value:%s", measure.MeasureType)
 	return key
 }
 
-func getKeyZaddTimestamp(measure entities.Measure) string {
+func getKeyZaddTimestamp(measure domain.Measure) string {
 	key := fmt.Sprintf("measure_timestamp:%s", measure.MeasureType)
 	return key
 }
